Refactor(evals): share mean/sum/round/rate via lib/stats.ts
Move report.ts's mean/sum/round/rate (the most complete copies) into
evals/lib/stats.ts and import them back. grade-suite.ts drops its local partial
mean/sum and replaces the ad hoc inline Math.round(...*10^n)/10^n rounding with
round()/sum() calls — numerically identical (round(x,n) === the prior inline
form), which also removes the rounding inconsistency the review flagged. Pure
refactor, no behavior change.
